Refactor code
Sort classes into packages
