+++ app/src/main/res/layout/fragment_chat.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".ChatFragment">
+    tools:context=".fragments.ChatFragment">
 
     <TextView
         android:id="@+id/fragment_chat_welcome_text"
