+++ app/src/main/res/layout/fragment_sign_in.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".SignInFragment">
+    tools:context=".fragments.SignInFragment">
 
     <TextView
         android:layout_width="wrap_content"
